+++ src/shared/config/config.ts
@@ -24,3 +24,4 @@ export const JWT_STORE_KEY = `${config.app.id}:token`
 export const LOCALE_STORAGE_KEY = `${config.app.id}:locale`
 export const USER_STORAGE_KEY = `${config.app.id}:user`
 export const ITEMS_TABLE_COLUMN_VISIBILITY_KEY = `${config.app.id}:items-table-column-visibility`
+export const ALL_ITEMS_TABLE_COLUMN_VISIBILITY_KEY = `${config.app.id}:all-items-table-column-visibility`