+++ src/shared/services/auth.interceptor.ts
@@ -1,6 +1,5 @@
-import type { InternalAxiosRequestConfig } from 'axios'
-
 import { JWT_STORE_KEY } from '@/shared/config/config'
+import type { InternalAxiosRequestConfig } from 'axios'
 
 export function authInterceptor(config: InternalAxiosRequestConfig): InternalAxiosRequestConfig {
   const token = localStorage.getItem(JWT_STORE_KEY)